+++ lib/src/codecs_base.dart
@@ -1,13 +1,15 @@
 // Copyright (c) 2023, Sudipto Chandra
 // All rights reserved. Check LICENSE file for details.
 
-export 'base2.dart';
 export 'base16.dart';
+export 'base2.dart';
 export 'base32.dart';
 export 'base64.dart';
-export 'codecs/base2.dart';
+export 'bigint.dart';
 export 'codecs/base16.dart';
+export 'codecs/base2.dart';
 export 'codecs/base32.dart';
 export 'codecs/base64.dart';
+export 'codecs/bigint.dart';
 export 'core/codec.dart';
 export 'core/converter.dart';